pyproject.toml
[tool.poetry]
name = "pycudahll"
version = "0.1.0"
description = ""
authors = ["Gabe Maayan <gabemgem@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.24.2"
hyperloglog = "^0.0.14"
cupy-cuda12x = "^12.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
